pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ollamacode"
version = "0.1.0"
description = "A CLI tool for AI-assisted coding using Ollama"
authors = [{name = "OllamaCode Team"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "click>=8.0",
    "requests>=2.28",
    "rich>=12.0",
    "gitpython>=3.1",
    "watchdog>=2.1",
    "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=22.0",
    "flake8>=5.0",
    "mypy>=1.0"
]

[project.scripts]
ollamacode = "ollamacode.cli:main"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true